requirements.txt
Django>=3.1
djangorestframework>=3.11
psycopg2>=2.7.5,<2.8.0
djangorestframework-simplejwt

flake8>=3.6.0,<3.7.0